--- a/crates/node_binding/src/raw_options/raw_module/mod.rs
+++ b/crates/node_binding/src/raw_options/raw_module/mod.rs
@@ -32,18 +32,10 @@ use crate::{JsFilename, ModuleObject, RawResolveOptions};
 /// and passed to rust side loader in [get_builtin_loader] when using with
 /// `builtin_loader`.
 #[napi(object)]
+#[derive(Debug)]
 pub struct RawModuleRuleUse {
   pub loader: String,
   pub options: Option<String>,
-}
-
-impl Debug for RawModuleRuleUse {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    f.debug_struct("RawModuleRuleUse")
-      .field("loader", &self.loader)
-      .field("options", &self.options)
-      .finish()
-  }
 }
 
 #[rspack_napi_macros::tagged_union]