--- a/lint-staged.config.mjs
+++ b/lint-staged.config.mjs
@@ -5,7 +5,8 @@ export default {
 	"*.{ts,tsx,js,cts,cjs,mts,mjs}": [
 		"pnpm run lint:js",
 		// ignore staged file list; execute x without extra args
-		() => "pnpm run x ae ci"
+		() => "pnpm run x ae update",
+		"git add ."
 	],
 	"package.json": "pnpm run check-dependency-version"
 };