--- a/lint-staged.config.mjs
+++ b/lint-staged.config.mjs
@@ -5,8 +5,7 @@ export default {
 	"*.{ts,tsx,js,cts,cjs,mts,mjs}": [
 		"pnpm run lint:js",
 		// ignore staged file list; execute x without extra args
-		() => "pnpm run x ae update",
-		"git add ."
+		() => "pnpm run x ae ci"
 	],
 	"package.json": "pnpm run check-dependency-version"
 };