--- a/packages/rspack-test-tools/tests/configCases/loader-parallel/additional-data-no-passthrough/a.js
+++ b/packages/rspack-test-tools/tests/configCases/loader-parallel/additional-data-no-passthrough/a.js
@@ -1,0 +1,1 @@
+module.exports = "a";