--- a/packages/rspack/src/config/adapter.ts
+++ b/packages/rspack/src/config/adapter.ts
@@ -264,8 +264,7 @@ const getRawModuleRule = (
 		rule.use = [
 			{
 				loader: rule.loader,
-				options: rule.options,
-				parallel: rule.parallel
+				options: rule.options
 			}
 		];
 	}