--- a/packages/rspack/src/config/adapterRuleUse.ts
+++ b/packages/rspack/src/config/adapterRuleUse.ts
@@ -87,7 +87,7 @@ export interface Diagnostic {
 	severity: "error" | "warning";
 }
 
-interface LoaderExperiments {
+export interface LoaderExperiments {
 	emitDiagnostic(diagnostic: Diagnostic): void;
 }
 
@@ -528,7 +528,7 @@ function resolveStringifyLoaders(
 	isBuiltin: boolean
 ) {
 	const obj = parsePathQueryFragment(use.loader);
-	let ident: string | null = null;
+	let ident = use.ident;
 
 	if (use.options === null) {
 	} else if (use.options === undefined) {
@@ -539,9 +539,21 @@ function resolveStringifyLoaders(
 	else if (typeof use.options === "object") obj.query = `??${(ident = path)}`;
 	else obj.query = `?${JSON.stringify(use.options)}`;
 
+	const parallelism = !!use.parallel;
+
+	if (parallelism && (!use.options || typeof use.options !== "object")) {
+		throw new Error(
+			`\`Rule.use.parallel\` requires \`Rule.use.options\` to be an object.\nHowever the received value is \`${use.options}\` under option path \`${path}\`\nInternally, parallelism is provided by passing \`Rule.use.ident\` to the loader as an identifier to ident the parallelism option\nYou can either replace the \`Rule.use.loader\` with \`Rule.use.options = {}\` or remove \`Rule.use.parallel\`.`
+		);
+	}
+
 	if (use.options && typeof use.options === "object") {
 		if (!ident) ident = "[[missing ident]]";
 		compiler.__internal__ruleSet.references.set(ident, use.options);
+		compiler.__internal__ruleSet.references.set(
+			`${ident}$$parallelism`,
+			parallelism
+		);
 		if (isBuiltin) {
 			compiler.__internal__ruleSet.builtinReferences.set(ident, use.options);
 		}