--- a/packages/rspack/src/config/defaults.ts
+++ b/packages/rspack/src/config/defaults.ts
@@ -248,6 +248,9 @@ const applyExperimentsDefaults = (
 
 	// IGNORE(experiments.parallelCodeSplitting): Rspack specific configuration for new code splitting algorithm
 	D(experiments, "parallelCodeSplitting", true);
+
+	// IGNORE(experiments.parallelLoader): Rspack specific configuration for parallel loader execution
+	D(experiments, "parallelLoader", false);
 };
 
 const applybundlerInfoDefaults = (